% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.Hybrid.Design.R
\name{Bayesian.Hybrid.Design}
\alias{Bayesian.Hybrid.Design}
\title{Bayesian Hybrid Design}
\usage{
Bayesian.Hybrid.Design(
  pt = 0.512,
  nt = 40,
  pc = 0.312,
  nc = 40,
  pch = 0.312,
  nche = 40,
  nch = 234,
  sig = 0.9,
  a0c = 0.001,
  b0c = 0.001,
  a0t = 0.001,
  b0t = 0.001,
  delta_threshold = 0.1,
  nsim = 10000,
  seed = 2000
)
}
\arguments{
\item{pt}{response rate for experimental arm in current study}

\item{nt}{number of patients in experimental arm in current study}

\item{pc}{response rate for control arm in current study}

\item{nc}{number of patients in control arm in current study}

\item{pch}{response rate for control treatment in historical study}

\item{nche}{Equivalent number of patients borrowed from historical study}

\item{nch}{Total number of patients in historical control}

\item{sig}{significance boundary defined as P(pt_hat > pc_hat|hybrid) > sig}

\item{a0c}{hyperprior for control response rate beta(a0c, b0c)}

\item{b0c}{hyperprior for control response rate beta(a0c, b0c)}

\item{a0t}{hyperprior for experimental response rate beta(a0t, b0t)}

\item{b0t}{hyperprior for experimental response rate beta(a0t, b0t)}

\item{delta_threshold}{Borrow when abs(pc_hat (current study) - pch) <= delta_threshold}

\item{nsim}{Number of replications to calculate power}

\item{seed=2000}{seed for simulations}
}
\value{
An object with values
\itemize{
\item power study power
\item delta.bound Minimum detectable difference of response rate in the current study using hybrid design
\item pc.bias bias estimate of posterior median ORR for control arm using hybrid design
\item pc.mse MSE of posterior median ORR for control arm using hybrid design
}
}
\description{
This function calculates the power and design parameters Bayesian Hybrid Design using dynamic power prior approach.
}
\examples{

Bayesian.Hybrid.Design(pt=0.512, nt=40,pc=0.312,nc=40,pch=0.312,nche=40,nch=234, sig=0.90,a0c=0.001,b0c=0.001,a0t=0.001,b0t=0.001,delta_threshold=0.1)

}
