% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{mixlink}
\alias{mixlink}
\title{Support of Distributions}
\usage{
mixlink(mix, x)
}
\description{
Returns the support of a distribution.
}
\details{
This function is exactly from the SAMprior R package version 1.1.1.
It is included to support the runSAM function in this package.

takes x and transforms it according to the defined link function of
the mixture
}
\keyword{internal}
