% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{support}
\alias{support}
\alias{mixlink}
\alias{mixinvlink}
\alias{mixJinv_orig}
\alias{mixlJinv_orig}
\alias{mixlJinv_link}
\alias{is.dlink}
\alias{is.dlink_identity}
\alias{is.mixidentity_link}
\alias{support.default}
\alias{support.betaMix}
\alias{support.gammaMix}
\alias{support.normMix}
\title{Support of Distributions}
\usage{
mixlink(mix, x)

mixinvlink(mix, x)

mixJinv_orig(mix, x)

mixlJinv_orig(mix, x)

mixlJinv_link(mix, l)

is.dlink(x)

is.dlink_identity(x)

is.mixidentity_link(mix, l)

support(mix)

\method{support}{default}(mix)

\method{support}{betaMix}(mix)

\method{support}{gammaMix}(mix)

\method{support}{normMix}(mix)
}
\arguments{
\item{mix}{Mixture distribution.}

\item{x}{Values to be transformed.}

\item{l}{Link values.}
}
\value{
A numeric vector of length 2 containing the lower and upper bounds
of the support for the specified mixture distribution. If the mixture
type is unknown (default method), it returns the character string
"Unknown mixture".
}
\description{
Returns the support of a distribution. This function is adapted from
the 'SAMprior' package (v1.1.1) to support the \code{\link{runSAM}} function.
}
\keyword{internal}
