% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.Hybrid.Analysis.R
\name{Bayesian.Hybrid.Analysis}
\alias{Bayesian.Hybrid.Analysis}
\title{Statistical Analysis for a Bayesian Hybrid Design}
\usage{
Bayesian.Hybrid.Analysis(
  Yt = 20,
  nt = 40,
  Yc = 12,
  nc = 40,
  Ych = 73,
  nche = 40,
  nch = 234,
  sig = 0.9,
  credlev = 0.8,
  a0c = 0.001,
  b0c = 0.001,
  a0t = 0.001,
  b0t = 0.001,
  delta_threshold = 0.1
)
}
\arguments{
\item{Yt}{Number of responses in experimental arm in current study}

\item{nt}{number of patients in experimental arm in current study}

\item{Yc}{Number of responses in control arm in current study}

\item{nc}{number of patients in control arm in current study}

\item{Ych}{Number of responses in control treatment in historical study}

\item{nche}{Equivalent number of patients borrowed from historical study}

\item{nch}{Total number of patients in historical control}

\item{sig}{significance boundary defined as P(pt_hat > pc_hat|hybrid) > sig}

\item{a0c}{hyperprior for control response rate beta(a0c, b0c)}

\item{b0c}{hyperprior for control response rate beta(a0c, b0c)}

\item{a0t}{hyperprior for experimental response rate beta(a0t, b0t)}

\item{b0t}{hyperprior for experimental response rate beta(a0t, b0t)}

\item{delta_threshold}{Borrow when abs(pc_hat (current study) - pch) <= delta_threshold}

\item{credlev:}{Credible interval level}
}
\value{
An object with values
\itemize{
\item prob.pt.gt.pc Probability of experimental arm having a better posterior response rate than control
\item median_hca Posterior median response rate for hybrid control
\item CI_hca Credible interval for median response rate for hybrid control
\item median_c Posterior median response rate for current study control
\item CI_c Credible interval for median response rate for current study control
\item median_t Posterior median response rate for current study experimental arm
\item CI_t Credible interval for median response rate for current experimental arm
\item delta.m Posterior median response rate difference for experimental arm vs control arm based on hybrid design
\item delta.CI Credible interval for posterior median response rate difference for experimental arm vs control arm  based on hybrid design
\item delta.m_trial Posterior median response rate difference for experimental arm vs control arm based on current study only
\item delta.CI_trial  Credible interval for posterior median response rate difference for experimental arm vs control arm based on current study only
\item conclusion Statistical inference conclusion for comparing the response rate of experimental arm vs control arm
}
}
\description{
This function performs the statistical analysis for a Bayesian Hybrid Design using dynamic power prior approach.
}
\examples{

Bayesian.Hybrid.Analysis(Yt=18, nt=40,Yc=13,nc=40,Ych=73,nche=40,nch=234)

}
