% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrow.wt.R
\name{borrow.wt}
\alias{borrow.wt}
\title{Borrowing weights from historical study}
\usage{
borrow.wt(
  Yc = 40 * 0.312,
  nc = 40,
  Ych = 234 * 0.312,
  nch = 234,
  nche = 40,
  a0c = 0.001,
  b0c = 0.001,
  delta_threshold = 0.1,
  method = "Empirical Bayes",
  theta = 0.5
)
}
\arguments{
\item{Yc}{number of responders for control arm in current study}

\item{nc}{number of patients in control arm in current study}

\item{Ych}{number of responders for control treatment in historical study}

\item{nch}{Total number of patients in historical control}

\item{nche}{Equivalent number of patients borrowed from historical study}

\item{a0c}{hyperprior for control response rate beta(a0c, b0c)}

\item{b0c}{hyperprior for control response rate beta(a0c, b0c)}

\item{delta_threshold}{Borrow when abs(pc (current study) - pch) <= delta_threshold}

\item{method}{Method for dynamic borrowing, "Empirical Bayes", "Bayesian p", "Generalized BC", "JSD"}

\item{theta}{A parameter with a range of (0, 1), and applicable to method = "Generalized BC" or "JSD"}
}
\value{
An object with values
\itemize{
\item a Global borrowing weight
\item wd Dynamic borrowing weight according to similarly of response rate
\item w Overall borrowing weight
}
}
\description{
This function calculates the borrowing weights for the hybrid control from historical study.
The following methods are included: (1) Empirical Bayes: dynamic weight is determined by maximizing marginal distribution;
(2) Bayesian P value: similarity measured by Bayesian P value comparing two distributions;
(3) Generalized Bhattacharyya Coefficient (BC): Bhattacharyya coefficient is a special case for theta = 0.5.
(4) Jensen-Shannon divergence (JSD): a measurement of similarity for two distributions.
}
\examples{

borrow.wt(Yc=40*0.312, nc=40, Ych=234*0.312, nch=234, nche=40, a0c=0.001, b0c=0.001, delta_threshold=0.1)

}
