% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrow.wt.R
\name{borrow.wt}
\alias{borrow.wt}
\title{Borrowing weights from historical study}
\usage{
borrow.wt(
  Yc,
  nc,
  Ych,
  nch,
  nche,
  a0c = 0.001,
  b0c = 0.001,
  delta_threshold = 0.1,
  method = "Empirical Bayes",
  theta = 0.5,
  eta = 1
)
}
\arguments{
\item{Yc}{A scalar number of responders for control arm in current study}

\item{nc}{A scalar number of patients in control arm in current study}

\item{Ych}{A scalar number of responders for control treatment in historical study}

\item{nch}{A scalar total number of patients in historical control}

\item{nche}{A scalar for the equivalent number of patients borrowed from historical study}

\item{a0c}{The 1st scale hyperprior hyperparameter for control response rate beta(a0c, b0c). Default 0.001.}

\item{b0c}{The 2nd hyperprior hyperparameter for control response rate beta(a0c, b0c). Default 0.001.}

\item{delta_threshold}{A scale threshold parameter. Only if abs(pc (current study) - pch) <= delta_threshold, we borrow from historical control. Default 0.1.}

\item{method}{A string characters. Method for dynamic borrowing, "Empirical Bayes", "Bayesian p", "Generalized BC", "JSD"}

\item{theta}{A scalar parameter with a range of (0, 1), and applicable to "Generalized BC". Default 0.5.}

\item{eta}{A scalar parameter with a range of (0, infty), and applicable to methods "Bayesian p", "Generalized BC", "JSD". Default 1.}
}
\value{
An object with values
\itemize{
\item a Global borrowing weight
\item wd Dynamic borrowing weight according to similarly of response rate
\item w Overall borrowing weight
}
}
\description{
This function calculates the borrowing weights for the hybrid control from historical study.
The following methods are included: (1) Empirical Bayes: dynamic weight is determined by maximizing marginal distribution;
(2) Bayesian P value: similarity measured by Bayesian P value comparing two distributions;
(3) Generalized Bhattacharyya Coefficient (BC): Bhattacharyya coefficient is a special case for theta = 0.5.
(4) Jensen-Shannon divergence (JSD): a measurement of similarity for two distributions.
}
\examples{

borrow.wt(Yc=40*0.3, nc=40, Ych=180*0.3, nch=180, nche=40, a0c=0.001, b0c=0.001, delta_threshold=0.1, eta=1)

}
