% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPP.analysis.R
\name{DPP.analysis}
\alias{DPP.analysis}
\title{Analysis of a Study Using Bayesian Hybrid Design using Dynamic Power Prior Framework}
\usage{
DPP.analysis(
  Yt = 39,
  nt = 60,
  Yc = 13,
  nc = 30,
  Ych = 90,
  nch = 200,
  nche = 30,
  a0c = 0.001,
  b0c = 0.001,
  a0t = 0.001,
  b0t = 0.001,
  delta_threshold = 0.1,
  method = "Empirical Bayes",
  theta = 0.5,
  eta = 1
)
}
\arguments{
\item{Yt}{A scalar. Response rate for experimental arm in current study}

\item{nt}{sample size for experimental arm}

\item{Yc}{A scalar. Response rate for control arm in current study}

\item{nc}{sample size for control arm}

\item{Ych}{A scalar. Randomization ratio for current study. r=1 means 1:1 and r=2 means 2:1.}

\item{nch}{A scalar. Specification of n_che in terms of times of nc; i.e. n_che = q*nc. Usually, q >= 1 and q <= nch/n_che.}

\item{nche}{maximum amount of borrowing in terms of equivalent number of subjects}

\item{a0c}{A scalar. Hyperprior for control response rate beta(a0c, b0c). Default 0.001.}

\item{b0c}{A scalar. Hyperprior for control response rate beta(a0c, b0c). Default 0.001.}

\item{a0t}{A scalar. Hyperprior for experimental response rate beta(a0t, b0t). Default 0.001.}

\item{b0t}{A scalar. Hyperprior for experimental response rate beta(a0t, b0t). Default 0.001.}

\item{delta_threshold}{Borrow when abs(pc_hat (current study) - pch) <= delta_threshold#'. Default 0.1.}

\item{method}{A string characters. Method for dynamic borrowing, "Empirical Bayes", "Bayesian p", "Generalized BC", "JSD". Default "Empirical Bayes".}

\item{theta}{A scalar parameter with a range of (0, 1), and applicable to method: "Generalized BC". Default 0.5.}

\item{eta}{A scalar parameter with a range of (0, infty), and applicable to method: "Bayesian p", "Generalized BC", "JSD". "Generalized BC" method requires two parameters theta and eta. Default 1.}
}
\value{
An object with values
\itemize{
\item w Borrowing weight
\item phat_pt_larger_pc P(ORR_trt > ORR_ctrl | data)
\item apost_c_trial Posterior distribution of ORR for concurrent control: Beta(apost_c_trial, bpost_c_trial)
\item bpost_c_trial Posterior distribution of ORR for concurrent control: Beta(apost_c_trial, bpost_c_trial)
\item apost_c_hca Posterior distribution of ORR for hybrid control: Beta(apost_c_hca, bpost_c_hca)
\item bpost_c_hca Posterior distribution of ORR for hybrid control: Beta(apost_c_hca, bpost_c_hca)
\item apost_t Posterior distribution of ORR for experimental arm: Beta(apost_t, apost_t)
\item bpost_t Posterior distribution of ORR for experimental arm: Beta(apost_t, apost_t)
}
}
\description{
This function Perform Analysis for a Study Using Bayesian Hybrid Design using Dynamic Power Prior Framework.
}
\examples{

o = DPP.analysis(Yt=39, nt=60, Yc=13, nc=30, Ych=90, nch=200, nche = 30, 
a0c= 0.001, b0c= 0.001, a0t= 0.001, b0t= 0.001,
delta_threshold = 0.1, method = "Empirical Bayes", 
theta = 0.5, eta = 1)

}
