% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactci.R
\name{exactci}
\alias{exactci}
\title{Clopper-Pearson Exact Confidence Interval}
\usage{
exactci(r, n, conflev)
}
\arguments{
\item{r}{Number of reponses}

\item{n}{Total number of subjects}

\item{conflev}{Confidence level. Always two-sided.}
}
\value{
Clopper-Pearson CI
}
\description{
This function calculates the Clopper-Pearson exact confidence interval
}
\examples{

exactci (r=4, n=20, conflev=0.95)

}
