% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.power.DPP.R
\name{explore.power.DPP}
\alias{explore.power.DPP}
\title{Power Calculation of Several power.DPP Function Parameters in Bayesian Hybrid Design}
\usage{
explore.power.DPP(
  method,
  pc,
  nc,
  pc.calib,
  q,
  delta,
  r,
  pch,
  nch,
  alpha = 0.1,
  a0c = 0.001,
  b0c = 0.001,
  a0t = 0.001,
  b0t = 0.001,
  delta_threshold = 0.1,
  theta = 0.5,
  eta = 1,
  nsim = 1e+05,
  seed = 2000,
  ncore = NULL
)
}
\arguments{
\item{method}{A vector of methods for dynamic borrowing for exploration, supports "Empirical Bayes", "Bayesian p", "Generalized BC", "JSD"}

\item{pc}{A vector of response rate for study control arm for exploration}

\item{nc}{A vector of sample size for study control arm for exploration}

\item{pc.calib}{A vector of response rate of study control arm for exploration. Each of these response rates is used for calibration such that the p-value equals to alpha (specified type I error) if pc = pc.calib}

\item{q}{A vector of ratio of historical control arm and the study control arm for exploration.}

\item{delta}{A scalar of difference between pt and pc to determine pt in each setting.}

\item{r}{A scalar of  ratio of study experimental arm and study control arm  for exploration}

\item{pch}{A scalar of the response rate for historical control arm}

\item{nch}{A scalar of the total number of patients in historical control}

\item{alpha}{A scalar. One sided type I error rate.}

\item{a0c}{A scalar. Hyperprior for control response rate beta(a0c, b0c). Default 0.001.}

\item{b0c}{A scalar. Hyperprior for control response rate beta(a0c, b0c). Default 0.001.}

\item{a0t}{A scalar. Hyperprior for experimental response rate beta(a0t, b0t). Default 0.001.}

\item{b0t}{A scalar. Hyperprior for experimental response rate beta(a0t, b0t). Default 0.001.}

\item{delta_threshold}{Borrow when abs(pc_hat (current study) - pch) <= delta_threshold#'. Default 0.1.}

\item{theta}{A scalar parameter with a range of (0, 1), and applicable to method: "Generalized BC". Default 0.5.}

\item{eta}{A scalar parameter with a range of (0, infty), and applicable to method: "Bayesian p", "Generalized BC", "JSD". "Generalized BC" method requires two parameters theta and eta. Default 1.}

\item{nsim}{A scalar. Number of replications to calculate power. Default 100,000.}

\item{seed}{A scalar. seed for simulations. Default 2000.}

\item{ncore}{description}
}
\value{
A dataframe containing the setting parameters and
\itemize{
\item type I error
\item power
\item delta boundary
\item mean PMD
\item sd(PMD)
}
}
\description{
This function calculates the power and design parameters Bayesian Hybrid Design using
parallel computing for efficiency. It is useful when there are multiple values of X
parameters to explore. If only one setting to calculate, please use the power.DPP function.
What is exploring, what parameters allow vectors for exploration.
}
\examples{

Res1 = explore.power.DPP(method=c("Empirical Bayes"),
pc=c(0.17,0.27,0.37),
nc=23:33,
pc.calib = 0.27,
q = c(1,1.5,2),
pch=0.27, nch=500,
alpha=0.1,
a0c=0.001, b0c=0.001, a0t=0.001, b0t=0.001,
delta_threshold=0.1,
theta=0.5, eta=1,
nsim = 1000, seed=2000,
ncore=16)

}
