% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.bound.R
\name{fisher.bound}
\alias{fisher.bound}
\title{Rejection boundary by Fisher's exact test}
\usage{
fisher.bound(pc, nc, nt, alpha = 0.1)
}
\arguments{
\item{pc}{response rate for control arm}

\item{nc}{number of patients in control arm}

\item{nt}{number of patients in experimental arm}

\item{alpha}{p-value threshold for significance. Alpha must be one-sided. Default 0.1.}
}
\value{
An object with values
\itemize{
\item M A 2-by-2 contingency table for response and treatment group
\item p p-value when the response rates just across the boundary
\item rc number of responders for control arm, determined by the response rate and sample size
\item nc sample size in control arm
\item rt response rate for experimental arm just acros the boundary
\item nt sample size in experimental arm
\item delta Minimum detectable difference of response rate between two groups
}
}
\description{
This function calculates the minimum detectable difference in response rate when using Fisher's exact test.
}
\examples{

fisher.bound(pc=0.3,nc=40,nt=40,alpha=0.1)

}
