% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.power.R
\name{fisher.power}
\alias{fisher.power}
\title{Power Calculation by Fisher's Exact Test}
\usage{
fisher.power(
  pt = 0.512,
  nt = 40,
  pc = 0.312,
  nc = 40,
  alpha = 0.05,
  nsim = 10000,
  seed = 2023
)
}
\arguments{
\item{pt}{Probablity of success in experimental arm}

\item{nt}{Number of subject in experimental arm}

\item{pc}{Probablity of success in control arm}

\item{nc}{Number of subject in control arm}

\item{alpha}{Type I error (one-sided)}

\item{nsim}{Number of replications in simulation}

\item{seed}{seed for simulation}
}
\value{
power
}
\description{
This function calculates the power by simulations using Fisher's exact test.
}
\examples{

fisher.power(pt = 0.512, nt = 40, pc = 0.312, nc = 40,nsim = 1000)

}
