% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fishers Exact Test for Response Rate in Two arms}
\usage{
fisher(Yc = 12, nc = 40, Yt = 19, nt = 40, alternative = "greater")
}
\arguments{
\item{r}{Number of reponses}

\item{n}{Total number of subjects}

\item{conflev}{Confidence level}
}
\value{
Fishers test object
}
\description{
This function performs Fishers exact test customized for comparing response rate for two arms.
}
\examples{

fisher(Yc=12, nc=40, Yt=19, nt=40)

}
