% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fishers Exact Test for Response Rate in Two arms}
\usage{
fisher(Yc, nc, Yt, nt, alternative = "greater")
}
\arguments{
\item{Yc}{A scalar. Number of subjects with response in experimental arm}

\item{nc}{A scalar. Number of subjects in control arm}

\item{Yt}{A scalar. Number of subjects with response  in control arm}

\item{nt}{A scalar. Number of subjects in experimental arm}
}
\value{
Fishers test object
}
\description{
This function performs Fishers exact test customized for comparing response rate for two arms.
}
\examples{

fisher(Yc=12, nc=40, Yt=19, nt=40)

}
