% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequentist.Analysis.R
\name{Frequentist.Analysis}
\alias{Frequentist.Analysis}
\title{Statistical Analysis Using Frequentist Method Based on Current Study Only}
\usage{
Frequentist.Analysis(Yt = 20, nt = 40, Yc = 12, nc = 40, conflev = 0.8)
}
\arguments{
\item{Yt}{Number of responses in experimental arm in current study}

\item{nt}{number of patients in experimental arm in current study}

\item{Yc}{Number of responses in control arm in current study}

\item{nc}{number of patients in control arm in current study}

\item{conflev}{Confidence Interval Level}
}
\value{
An object with values
\itemize{
\item pt Response rate of experimental arm
\item pc Response rate of control arm
\item delta Difference of response rate
\item exactCI.c Clopper-Pearson CI for control arm
\item exactCI.t Clopper-Pearson CI for experimental arm
\item p.fisher Fishers exact test p-value. Always one-sided.
}
}
\description{
This function performs the statistical analysis using frequentist method based on current study only.
}
\examples{

Frequentist.Analysis(Yt=18, nt=40,Yc=13, nc=40, conflev=0.8)

}
