% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDPP.R
\name{plotDPP}
\alias{plotDPP}
\title{Plot of DPP Object produced by DPP.analysis() function.}
\usage{
plotDPP(DPP)
}
\arguments{
\item{DPP}{DPP object produced by DPP.analysis() function.}
}
\description{
Plot the posterior distributions of OR in control arm and experimental arm. DPP object produced by DPP.analysis() function.
}
\details{
This function Perform Analysis for a Study Using Bayesian Hybrid Design using Dynamic Power Prior Framework.
}
\examples{

o = DPP.analysis(Yt=39, nt=60, Yc=13, nc=30, Ych=90, nch=200, nche = 30,
a0c= 0.001, b0c= 0.001, a0t= 0.001, b0t= 0.001,
delta_threshold = 0.1, method = "Empirical Bayes",
theta = 0.5, eta = 1)

plot.DPP(DPP=o)

}
