% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMD.R
\name{plotPMD}
\alias{plotPMD}
\title{Plot of Posterior Mean Difference (PMD) Object produced by power.DPP() function.}
\usage{
plotPMD(o, o2 = NULL)
}
\arguments{
\item{o}{Object produced by power.DPP() function.}

\item{o2}{Second object produced by power.DPP() function. Default NULL.}
}
\value{
An object with values
\itemize{
\item mean_PMD Mean of Posterior Mean Difference (PMD), calculated by the simulations.
\item sd_PMD Standard Deviation of Posterior Mean Difference (PMD), calculated by the simulations.
\item CI95_PMD 95 percent credible interval for PMD, calculated by the simulations.
}
}
\description{
Plot the distribution of posterior mean difference (PMD) by simulations.
}
\details{
This function Perform Analysis for a Study Using Bayesian Hybrid Design using Dynamic Power Prior Framework.
}
\examples{

o1a=power.DPP(pt=0.65, nt=60,pc=0.45,nc=30, pc.calib = 0.45, pch=0.45,nche=60,nch=200, alpha=0.1,a0c=0.001,b0c=0.001,a0t=0.001,b0t=0.001,delta_threshold=0.1, nsim=1000)
o1b=power.DPP(pt=0.65, nt=60,pc=0.45,nc=30, pc.calib = 0.45, pch=0.45,nche=30,nch=200, alpha=0.1,a0c=0.001,b0c=0.001,a0t=0.001,b0t=0.001,delta_threshold=0.1, nsim=1000)
plotPMD(o=o1a, o2=o1b) #perfect
plotPMD(o=o1a) #perfect

}
