% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.binom.test.R
\name{power.binom.test}
\alias{power.binom.test}
\title{Power calculation using exact binomial test for one sample problem}
\usage{
power.binom.test(
  n = 20,
  p = 0.5,
  p0 = 0.3,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  nsim = 1e+05,
  seed = 2025
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Assumed proportion of experimental treatment}

\item{p0}{Historical proportion as a benchmark}

\item{alpha}{Type I error}

\item{alternative}{Alternative hypothesis: "two.sided", "less", "greater".}

\item{nsim}{Number of simulated trials, default 100,000}

\item{seed}{Seed for simulation}
}
\value{
Power
}
\description{
This function performs Fishers exact test customized for comparing response rate for two arms.
}
\examples{

power.binom.test (n=110, p=0.4, p0=0.128)

}
