% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.DPP.R
\name{power.DPP}
\alias{power.DPP}
\title{Bayesian Hybrid Design}
\usage{
power.DPP(
  pt,
  nt,
  pc,
  nc,
  pc.calib,
  pch,
  nche,
  nch,
  alpha = 0.1,
  tau = NULL,
  a0c = 0.001,
  b0c = 0.001,
  a0t = 0.001,
  b0t = 0.001,
  delta_threshold = 0.1,
  method = "Empirical Bayes",
  theta = 0.5,
  eta = 1,
  datamat = NULL,
  w0 = NULL,
  nsim = 1e+05,
  seed = 2000
)
}
\arguments{
\item{pt}{response rate for experimental arm in current study}

\item{nt}{number of patients in experimental arm in current study}

\item{pc}{Response rate for control arm in current study}

\item{nc}{number of patients in control arm in current study}

\item{pc.calib}{Required for calibration if tau is not provided. Response rate for control arm in current study for calibration. Usually, pc.calib = pch.}

\item{pch}{Response rate for control treatment in historical study}

\item{nche}{Equivalent number of patients borrowed from historical study}

\item{nch}{Total number of patients in historical control}

\item{alpha}{A scalar. One sided type I error rate. Required for calibration if tau is not provided.}

\item{tau}{Calibrated threshold for statistical significance. If tau is not provided, it will be calculated by calibration to type I error alpha.}

\item{a0c}{Hyperprior for control response rate beta(a0c, b0c)}

\item{b0c}{Hyperprior for control response rate beta(a0c, b0c)}

\item{a0t}{Hyperprior for experimental response rate beta(a0t, b0t)}

\item{b0t}{Hyperprior for experimental response rate beta(a0t, b0t)}

\item{delta_threshold}{Borrow when abs(pc_hat (current study) - pch) <= delta_threshold}

\item{method}{Method for dynamic borrowing, "Empirical Bayes", "Bayesian p", "Generalized BC", "JSD"}

\item{theta}{A parameter with a range of (0, 1), and applicable to method: "Generalized BC".}

\item{eta}{A parameter with a range of (0, infty), and applicable to method: "Bayesian p", "Generalized BC", "JSD". "Generalized BC" method requires two parameters theta and eta.}

\item{datamat}{A matrix with dimension nsim * 2 containing the pre-simulated data for the study treatment (1st column) and control  (1st column) groups, respectively. If not supplied, binomial random Monte Carlo samples will be generated in the function.}

\item{w0}{prior power parameters w. If not specified (default), w_d is calculated by the specified method for dynamic borrowing.}

\item{nsim}{Number of replications to calculate power}

\item{seed=2000}{seed for simulations}
}
\value{
An object with values
\itemize{
\item power study power
\item delta.bound Minimum detectable difference of response rate in the current study using hybrid design
\item pc.PMD Posterior mean difference between hybrid control and concurrent control
\item pc.sd  sd of PMD
}
}
\description{
This function calculates the power and design parameters Bayesian Hybrid Design using dynamic power prior approach.
}
\examples{

o=power.DPP(pt=0.5, nt=40,pc=0.3,nc=40, pc.calib = 0.3, pch=0.3,nche=40,nch=180, alpha=0.1,a0c=0.001,b0c=0.001,a0t=0.001,b0t=0.001,delta_threshold=0.1)

}
