% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.Prop.Test.Sample.Size.R
\name{Two.Prop.Test.Sample.Size}
\alias{Two.Prop.Test.Sample.Size}
\title{Sample Size Calibration for Two-Sample Proportions}
\usage{
Two.Prop.Test.Sample.Size(p1, p2, alpha, beta, r)
}
\arguments{
\item{p1}{Arm 1 proportion (control)}

\item{p2}{Arm 2 proportion (experimental)}

\item{alpha}{Type I error rate (one-sided)}

\item{beta}{type II error rate (1-power)}

\item{r}{Ratio of Arm 2 (experimental) vs Arm 1 (control)}
}
\value{
sample size for arm 1 (control) and total
}
\description{
This function calculates the sample size for comparing two-sample proportions
https://www2.ccrb.cuhk.edu.hk/stat/proportion/Casagrande.htm
}
\details{
p1,p2,alpha,beta,r
}
\examples{

Two.Prop.Test.Sample.Size (p1=0.27, p2=0.47, alpha = 0.10, beta=0.2, r=2)

}
